pyproject.toml
[project]
name = "multi-agentes-question"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.115.14",
    "ipython>=9.4.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "mysql-connector-python>=9.3.0",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt>=2.10.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "tzdata>=2025.2",
]
